move out of progress_common to progress_async where possible
