+++ archive/progress_async/src/test/scala/uk/ac/wellcome/platform/archive/progress_async/fixtures/ProgressAsyncFixture.scala
@@ -1,5 +1,7 @@
 package uk.ac.wellcome.platform.archive.progress_async.fixtures
 
+import akka.NotUsed
+import akka.stream.scaladsl.Flow
 import com.google.inject.Guice
 import uk.ac.wellcome.messaging.test.fixtures.Messaging
 import uk.ac.wellcome.messaging.test.fixtures.SNS.Topic
@@ -7,16 +9,13 @@ import uk.ac.wellcome.messaging.test.fixtures.SQS.QueuePair
 import uk.ac.wellcome.platform.archive.common.fixtures.RandomThings
 import uk.ac.wellcome.platform.archive.common.modules._
 import uk.ac.wellcome.platform.archive.common.progress.fixtures.ProgressTrackerFixture
-import uk.ac.wellcome.platform.archive.common.progress.models.Progress
+import uk.ac.wellcome.platform.archive.common.progress.flows.ProgressUpdateFlow
+import uk.ac.wellcome.platform.archive.common.progress.models.{Progress, ProgressUpdate}
 import uk.ac.wellcome.platform.archive.common.progress.modules.ProgressTrackerModule
 import uk.ac.wellcome.platform.archive.common.progress.monitor.ProgressTracker
-import uk.ac.wellcome.platform.archive.progress_async.modules.{
-  ConfigModule,
-  TestAppConfigModule
-}
-import uk.ac.wellcome.platform.archive.progress_async.{
-  ProgressAsync => ProgressApp
-}
+import uk.ac.wellcome.platform.archive.progress_async.modules.{ConfigModule, TestAppConfigModule}
+import uk.ac.wellcome.platform.archive.progress_async.{ProgressAsync => ProgressApp}
+import uk.ac.wellcome.storage.dynamo.DynamoConfig
 import uk.ac.wellcome.storage.fixtures.LocalDynamoDb.Table
 import uk.ac.wellcome.storage.fixtures.{LocalDynamoDb, S3}
 import uk.ac.wellcome.test.fixtures.TestWith
@@ -33,6 +32,20 @@ trait ProgressAsyncFixture
     val createdProgress = createProgress()
     val storedProgress = monitor.initialise(createdProgress)
     testWith(storedProgress)
+  }
+
+  def withProgressUpdateFlow[R](table: Table)(
+    testWith: TestWith[(
+      Flow[ProgressUpdate, Progress, NotUsed],
+        ProgressTracker
+      ),
+      R]): R = {
+
+    val progressTracker = new ProgressTracker(
+      dynamoDbClient,
+      DynamoConfig(table = table.name, index = table.index)
+    )
+    testWith((ProgressUpdateFlow(progressTracker), progressTracker))
   }
 
   def withApp[R](qPair: QueuePair, topic: Topic, table: Table)(
